refactor: centralize InputsSnapshot type
